Proposed inclusion under the Narrative section
While the rationale is helpful for understanding why section.text is constrained, the initial request for more guidance that came from one of our implementing jurisdictions was the result of solution vendors asking for clarity on what content IPS expected to show up in the required narrative field (pointing to the base guidance that indicates inclusion of "clinically relevant information" is ambiguous).

We're proposing the following sub-section that provides implementers with reasoning behind why IPS doesn't currently dictate the actual content and points towards emerging best practices for implementers & vendors who are considering what to include in the section narrative.